requirements/default.txt
celery>=3.1.0; python_version<"3.7"
celery>=4.3.0; python_version>="3.7"
tornado>=5.0.0,<6.0.0; python_version<"3.5.2"
tornado>=5.0.0,<7.0.0; python_version>="3.5.2"
humanize==0.5.1
python-influxdb==5.2.3
pytz